fix stuuuuupid bug in ADSR implementation
This bug meant ADSR returned an undefined value for one tick while
transitioning from decay to sustain. On linux and osx this was never
noticed, but on windows it always popped out as a NaN, which screwed
up the filter feedback from then on, so all sounds were being cut
off early!